Introduce column separator items; change most of ChatView to use qreal
These items will be used to drag the columns; also they will display a fancy
vertical line at some point (this FR is years old!).

Also changed most (or all?) our geometry-related values to qreal rather than int,
since QGV operates on qreals. Dunno if that is less efficient though; needs evaluation.

+++ src/qtui/chatview.h
@@ -57,7 +57,7 @@ class ChatView : public QGraphicsView, public AbstractChatView {
     virtual void resizeEvent(QResizeEvent *event);
 
   protected slots:
-    virtual void sceneHeightChanged(int height);
+    virtual void sceneHeightChanged(qreal height);
 
   private:
     ChatScene *_scene;
